Move to vitest for realm server tests, point default commands to it
