+++ packages/realm-server/tests-vitest/scripts/run-vitest-with-test-pg.sh
@@ -19,4 +19,4 @@ NODE_NO_WARNINGS=1 \
 PGPORT=55436 \
 STRIPE_WEBHOOK_SECRET=stripe-webhook-secret \
 STRIPE_API_KEY=stripe-api-key \
-qunit --require ts-node/register/transpile-only "$@" tests/index.ts
+vitest run --silent $@
